Clean up/random bugfixes

+++ app/src/main/java/com/ajzia/simplist/screens/products/CategoryDisplay.kt
@@ -9,7 +9,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ajzia.simplist.model.Product
 
@@ -37,7 +36,7 @@ fun CategoryDisplay(
       onAdd = onAdd
     )
 
-    if (isVisible) {
+    if (isVisible && products.isNotEmpty()) {
       ProductGrid(
         products = products,
         onRemove = { onRemove(it) }
